arreglos varios

+++ backend/main/resources/Calificacion.py
@@ -2,21 +2,27 @@ from flask_restful import Resource
 from flask import request,jsonify
 from .. import db
 from main.models import CalificacionModel
+from flask_jwt_extended import jwt_required,get_jwt_identity,get_jwt
 
 
 #Recurso Calificacion
 class Calificacion(Resource):
-   
+    @jwt_required(optional=True)
     def get(self, id):
         calificacion = db.session.query(CalificacionModel).get_or_404(id)
         return calificacion.to_json_complete()
     
+    @jwt_required()
     def delete(self, id):
-        
+        usuario_id = get_jwt_identity()
         calificacion = db.session.query(CalificacionModel).get_or_404(id)
-        db.session.delete(calificacion)
-        db.session.commit()
-        return '', 204
+        claims = get_jwt()
+        if calificacion.usuarioId == usuario_id or claims['admin']: #verifico si el usuario es el dueño de la calificacion o si es el admin y puedo proceder a eliminarla
+            db.session.delete(calificacion)
+            db.session.commit()
+            return 'eliminacion exitosa', 204
+        else:
+            return 'permiso denegado',403
 
 
     #Modificar recurso
@@ -33,6 +39,7 @@ class Calificacion(Resource):
 #Recurso Calificaciones
 class Calificaciones(Resource):
     #Obtener lista de recursos
+    @jwt_required(optional=True)
     def get(self):
         page=1
         per_page=10
